Add files via upload
This is an applet program which enables you to create a node just by clicking the mouse wherever you wish to have it.
In the top 3 boxes you have to enter the edge details i.e(node1,node2,edgeweight) and click submit. It is not necessary to enter the details of an edge with weight as 0. Then next two text boxes before "FindShortestPath" Button are to enter the details of source node and destination nodes for which we are intrested in finding the shortest path.